fixed changed logon procedure

+++ postnl.sh
@@ -1,21 +1,21 @@
 #!/bin/sh
 
+#===================================================================================================================================================================
+# This script can be used to get a json with Track & Trace of packages of Postnl.
+# You need a postnl account to use this script
+#
+# Version: 1.3  - glsf91 - 4-7-2021"
+#
+# DON'T RUN this script to many times in a hour. Use at our own risk."
+#==================================================================================================================================================================="
+
+
 # Change USER SETTINGS below if needed
 
-#If using proxy uncomment next line and change ipaddress and port
-#PROXY="--proxy 192.168.1.2:8888"
+#PROXY="--proxy 192.168.1.97:8888"   #If using proxy uncomment line and change ipaddress and port
 
 # end USER SETTINGS
 
-# "==================================================================================================================================================================="
-# "This script can be used to get a json with Track & Trace of packages of Postnl."
-# "You need a postnl account to use this script"
-# 
-# "Version: 1.2  - glsf91 - 1-5-2020"
-#
-# "DON'T RUN this script to many times in a hour. Use at our own risk."
-# "==================================================================================================================================================================="
-# ""
 
 usage() {
         echo ""
@@ -27,12 +27,20 @@ usage() {
         -u <username> (mandatory)
         -p <password> (mandatory)
         -d <directory for temporary files> (mandatory)
-		-f Force new login (optional)
+        -f Force new login (optional)
+        -v with debug output (optional)
         -h Display this help text
 		
         example:
             -u \"myusername\" -p \"mypassword\" -d \"/tmp/postnl\"
         "
+}
+
+debug_output() {
+	if $DEBUG
+	then 
+		echo "$*"; 
+	fi
 }
 
 exit_abnormal() {                              # Function: Exit with error.
@@ -56,7 +64,7 @@ readInbox() {
 
 	ACCODE=`cat $1`
 
-	RESPFILE="$TMPDIR/postnl-resp5.tmp"
+	RESPFILE="$TMPDIR/postnl-resp6.tmp"
 	removeFile $RESPFILE
 
 	STATUSCODE=`curl $PROXY -w "\n%{http_code}" -k -o $RESPFILE -b $COOKIEFILE -c $COOKIEFILE -s -H "Authorization: Bearer $ACCODE" "https://jouw.postnl.nl/web/api/default/inbox"`
@@ -73,7 +81,7 @@ readInbox() {
 
 	if [ -s $RESPFILE ]
 	then
-		cp $RESPFILE $INBOXFILE
+		mv $RESPFILE $INBOXFILE
 		echo "Succeeded: inbox json received from postnl"
 		echo "`date '+%Y-%m-%d %H:%M:%S'`" > $TMPDIR/lastupdate.log
 	else
@@ -88,28 +96,28 @@ USERNAME=""
 PASSWORD=""
 TMPDIR=""
 FORCELOGIN=false
+DEBUG=false
 
 PROGARGS="$@"
 
 #get command options
-while getopts "u:p:d:fh" opt $PROGARGS
+while getopts "u:p:d:fhv" opt $PROGARGS
 do
 	case $opt in
 		u)
-#			echo "Set username: $OPTARG"
 			USERNAME=$OPTARG
 			;;
 		p)
-#			echo "Set password: $OPTARG"
 			PASSWORD=$OPTARG
 			;;
 		d)
-#			echo "Set directory for temporary files: $OPTARG"
 			TMPDIR=$OPTARG
 			;;
 		f)
-#			echo "Set force login"
 			FORCELOGIN=true
+			;;
+		v)
+			DEBUG=true
 			;;
 		h)  usage
 			exit 1
@@ -127,6 +135,14 @@ do
 			;;
 	esac
 done
+
+
+debug_output "Set username: $USERNAME"
+debug_output "Set password: $PASSWORD"
+debug_output "Set directory for temporary files: $TMPDIR"
+debug_output "Set debug output: $DEBUG"
+debug_output "Set force login:  $FORCELOGIN"
+
 
 # Check 
 [ -z "$USERNAME" ] && exit_abnormal
@@ -172,15 +188,35 @@ removeFile $COOKIEFILE
 
 # Step 1
 
+debug_output ""
+debug_output "Start step 1"
+
 REQUESTVERIFICATIONTOKEN=""
+RETURNURL=""
 
 RESPFILE="$TMPDIR/postnl-resp1.tmp"
 removeFile $RESPFILE
 
-STATUSCODE=`curl $PROXY -w "\n%{http_code}" -o $RESPFILE -k -s -b non-existing -c $COOKIEFILE https://jouw.postnl.nl/identity/Account/Login`
-if [ ! $STATUSCODE = "200" ]
+
+# Generate CodeVerifier, stateValue and codeChallenge
+
+RANDOM=`dd bs=32 count=1 status=none </dev/urandom`
+CODEVERIFIER=`echo -n $RANDOM | hexdump -e '32/1 "%02x""\n"'`
+debug_output "CodeVerifier: $CODEVERIFIER"
+TMP1=`echo -n "$CODEVERIFIER" | openssl dgst -sha256 -binary | openssl base64`
+CODECHALLENGE=`echo -n $TMP1 | sed -e 's/\+/-/g' -e 's/\//_/g' -e 's/=//g'`
+debug_output "codeChallenge: $CODECHALLENGE"
+
+STATEVALUE=`dd bs=32 count=1 </dev/urandom status=none | hexdump -e '32/1 "%02x""\n"'`
+debug_output "stateValue: $STATEVALUE"
+
+OUTPUTCURL=`curl --tlsv1.2 $PROXY -i -o $RESPFILE -k -b $COOKIEFILE -c $COOKIEFILE -s -L -w %{url_effective} --request GET --header 'Connection: close' "https://jouw.postnl.nl/identity/connect/authorize?client_id=poa-profiles-web&audience=poa-profiles-api&scope=openid%20profile%20email%20poa-profiles-api%20hashed-data&response_type=code&code_challenge_method=S256&code_challenge=$CODECHALLENGE&prompt=prompt&state=$STATEVALUE&redirect_uri=https://jouw.postnl.nl/account/login&ui_locales=nl_NL"`
+#debug_output "Output identity/connect/authorize: $OUTPUTCURL"
+
+if ! echo $OUTPUTCURL | grep -q "^https://jouw.postnl.nl"
 then
-	echo "Error: Wrong responsecode $STATUSCODE received from Account/Login step 1 request. See $RESPFILE. Aborting."
+	echo "Error: Wrong redirect received from connect/authorize request. See $RESPFILE. Aborting."
+	echo "       Response: $OUTPUTCURL"
 	exit 1
 fi
 
@@ -191,33 +227,48 @@ then
 	exit 1
 fi
 
-# echo "RequestVerificationToken: $REQUESTVERIFICATIONTOKEN"
+debug_output "RequestVerificationToken: $REQUESTVERIFICATIONTOKEN"
 
-URLWITHSTATIC=`cat $RESPFILE | tr '\n' '\r' | sed "s/.*src=\"\([^\"]*\).*/\1/"` 
-if [ ${#URLWITHSTATIC} -lt 35 ] || [ ${#URLWITHSTATIC} -gt 50 ]
+URLWITHSTATIC=`cat $RESPFILE | tr '\n' '\r' | sed "s/.*script.type=.*src=\"\([^\"]*\).*/\1/"` 
+if [ ${#URLWITHSTATIC} -lt 35 ] || [ ${#URLWITHSTATIC} -gt 90 ]
 then 
 	echo "Error: UrlWithStatic has wrong size. Aborting."
 	exit 1
 fi
 
-echo "UrlWithStatic: $URLWITHSTATIC"
+debug_output "UrlWithStatic: $URLWITHSTATIC"
+
+
+RETURNURL=`cat $RESPFILE | tr '\n' '\r' | sed "s/.*id=\"ReturnUrl\" name=\"ReturnUrl\" value=\"\([^\"]*\).*/\1/"` 
+if [ ${#RETURNURL} -lt 200 ] || [ ${#RETURNURL} -gt 600 ]
+then 
+	echo "Error: ReturnUrl has wrong size. Aborting."
+	exit 1
+fi
+
+RETURNURL=`echo "$RETURNURL"| sed 's/\&amp;/\&/g'`
+
+debug_output "ReturnUrl: $RETURNURL"
 
 
 # Step 2
+debug_output ""
+debug_output "Start step 2"
 
 RESPFILE="$TMPDIR/postnl-resp2.tmp"
 removeFile $RESPFILE
 
 RANDOMSENSORDATA=`dd bs=11 count=1 status=none </dev/urandom | hexdump -e '11/1 "%02x""\n"'`
+debug_output "RANDOMSENSORDATA: $RANDOMSENSORDATA"
 
-STATUSCODE=`curl $PROXY -w "\n%{http_code}" -o $RESPFILE -k -s -b $COOKIEFILE -c $COOKIEFILE --request POST --header 'Expect:' --header 'Content-Type: text/plain' --data-raw '{"sensor_data":"'$RANDOMSENSORDATA'1.66-1,2,-94,-100,Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.163 Safari/537.36,uaend,12147,20030107,en-US,Gecko,3,0,0,0,390320,2639919,1920,1160,1920,1200,1778,965,1794,,cpen:0,i1:0,dm:0,cwen:0,non:1,opc:0,fc:0,sc:0,wrc:1,isc:0,vib:1,bat:1,x11:0,x12:1,8334,0.517452032258,793181319959.5,loc:-1,2,-94,-101,do_en,dm_en,t_en-1,2,-94,-105,0,-1,0,0,864,832,0;1,0,0,0,883,851,0;-1,2,-94,-102,0,-1,0,0,864,832,0;1,0,0,0,883,851,0;-1,2,-94,-108,-1,2,-94,-110,-1,2,-94,-117,-1,2,-94,-111,-1,2,-94,-109,-1,2,-94,-114,-1,2,-94,-103,-1,2,-94,-112,https://jouw.postnl.nl/identity/Account/Login-1,2,-94,-115,1,32,32,0,0,0,0,1,0,1586362639919,-999999,16970,0,0,2828,0,0,3,0,0,B5B13FCDF4CA36C7D95DCC2A23DE191A~-1~YAAQDk5lX22iFDpxAQAApeOUWgO7v0AjsdlFKoX+6HQsg8w34qbHV5uqhWQDQ+mBZtaqUOQ0dH9nvZCDqxUeua/VxcrKJV6UEnpiVKtc/toBhuGBK5H95xDMtfB4HxcBesbqctyj1TsB5Kdg7muJ3P7sWOR/49n4c5O3gO74zdQWhHPGQpPI81z7vynNzrsagRpaH+6eojAeufTm/uzqGq7QZR/eoXe/iN6yPmCR9s6u+3twyemZty27C4vuyhbAdDvQDoe65qof6ADb2APRVSqEMUnn3tiT/kFHgKR39U2QFZQN1gVJLLU=~-1~-1~-1,29634,-1,-1,30261693-1,2,-94,-106,0,0-1,2,-94,-119,-1-1,2,-94,-122,0,0,0,0,1,0,0-1,2,-94,-123,-1,2,-94,-124,-1,2,-94,-126,-1,2,-94,-127,-1,2,-94,-70,-1-1,2,-94,-80,94-1,2,-94,-116,2639947-1,2,-94,-118,79837-1,2,-94,-121,;4;-1;0"}'  https://jouw.postnl.nl$URLWITHSTATIC`
+STATUSCODE=`curl --tlsv1.2 $PROXY -w "\n%{http_code}" -o $RESPFILE -k -s -b $COOKIEFILE -c $COOKIEFILE --request POST --header 'Expect:' --header 'Connection: close' --header 'Content-Type: text/plain' --data-raw '{"sensor_data":"'$RANDOMSENSORDATA'1.7-1,2,-94,-100,Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36,uaend,12147,20030107,en-US,Gecko,3,0,0,0,399909,5876996,1920,1160,1920,1200,1751,885,1767,,cpen:0,i1:0,dm:0,cwen:0,non:1,opc:0,fc:0,sc:0,wrc:1,isc:0,vib:1,bat:1,x11:0,x12:1,8323,0.19290546196,812667938498,0,loc:-1,2,-94,-101,do_en,dm_en,t_en-1,2,-94,-105,-1,2,-94,-102,-1,2,-94,-108,-1,2,-94,-110,-1,2,-94,-117,-1,2,-94,-111,0,99,-1,-1,-1;-1,2,-94,-109,0,99,-1,-1,-1,-1,-1,-1,-1,-1,-1;-1,2,-94,-114,-1,2,-94,-103,-1,2,-94,-112,https://jouw.postnl.nl/account/en-GB/login-1,2,-94,-115,1,32,32,99,99,0,198,516,0,1625335876996,9,17387,0,0,2897,0,0,517,198,0,ECB30B14F9D32567F2245B547EED3CF7~-1~YAAQB05lX6pbcAN6AQAA206RbQZjqi0USY2vjHf9GR5KyECMd51xAhhIy3vpemr52YzYtIFlsn1yTljV4TYpuD3dvDcjFeaJsK4z/0U0KyccaLCMWfZne8W2Ghezs3aWZJUSdZZJ284ZRNJSyFgEM1kGqk3beh5cn2ibAt4qH/aDnZTMrHUBKI23jGw73WS7FEqPnJNP03mJNDEQhdBe124QdK/I721ZBtgi2wpLr06NWJtH5EMY83fuz4hkWpHZfXFywZwv+cFALAZdD5g2hmWLbmNsx7X+finaD+jAJvfrgwjR2iVIDiz5mahMX8dkorZdN3++N0rBOia4TZKsdLXD4AheAPL/B4shFsB103sRvF3r6LsJoxlAQUhhYCLVJhZ2M4cGCZ9w~-1~||1-BvHxpIrdVU-1-10-1000-2||~-1,37993,760,-395545193,30261693,PiZtE,20316,37,0,-1-1,2,-94,-106,9,2-1,2,-94,-119,20,20,20,20,40,20,0,0,0,0,0,0,20,120,-1,2,-94,-122,0,0,0,0,1,0,0-1,2,-94,-123,-1,2,-94,-124,0.3594e66556ef8,0.db2eef4a7fb0d,0.74a11d3436c6e,0.d38e92cbbc00a,0.90a5d01022534,0.755e18918bb8d,0.0b87885a41aee,0.26f8c6a97a988,0.1e751ac6043ce,0.7ca004bb7aab7;0,0,0,0,0,2,0,5,0,2;0,0,0,2,1,14,2,31,0,8;ECB30B14F9D32567F2245B547EED3CF7,1625335876996,BvHxpIrdVU,ECB30B14F9D32567F2245B547EED3CF71625335876996BvHxpIrdVU,1,1,0.3594e66556ef8,ECB30B14F9D32567F2245B547EED3CF71625335876996BvHxpIrdVU10.3594e66556ef8,180,209,77,110,220,153,52,51,72,117,167,127,60,109,224,158,212,154,204,219,200,130,21,163,249,213,88,150,185,133,79,72,205,0,1625335877266;-1,2,-94,-126,-1,2,-94,-127,11321144241322243122-1,2,-94,-70,-739578230;-1395479418;dis;,7,8;true;true;true;-120;true;24;24;true;false;-1-1,2,-94,-80,5636-1,2,-94,-116,17630982-1,2,-94,-118,123173-1,2,-94,-129,ef9216fc326496a8227a7fc12bd42b7add55e95003515406b71315a7378f5030,1,59e04af447a7870fd61e1342399b41011fd7568b0419030175f3efbee83ca251,Google Inc. (NVIDIA),ANGLE (NVIDIA, NVIDIA GeForce GTX 1650 SUPER Direct3D11 vs_5_0 ps_5_0, D3D11-27.21.14.5671),95f5b71fe531f867faa814bdd4050dd8057206d53ecec1163523560525884870,33-1,2,-94,-121,;3;4;0"}'  https://jouw.postnl.nl$URLWITHSTATIC`
 if [ ! $STATUSCODE = "201" ]
 then
 	echo "Error: Wrong responsecode $STATUSCODE received from static step 2 request. See $RESPFILE. Aborting."
 	exit 1
 fi
 
-if ! grep -q "{\"success\":true}" $RESPFILE
+if ! grep -q "{\"success\": true}" $RESPFILE
 then
 	echo "Error: No success: true received from static request. See $RESPFILE. Aborting."
 	exit 1
@@ -225,11 +276,14 @@ fi
 
 
 # Step 3
+debug_output ""
+debug_output "Start step 3"
 
 RESPFILE="$TMPDIR/postnl-resp3.tmp"
 removeFile $RESPFILE
 
-STATUSCODE=`curl $PROXY -i -w "\n%{http_code}" -o $RESPFILE -k -s -b $COOKIEFILE -c $COOKIEFILE -L -d "ReturnUrl=&Username=$USERNAME&Password=$PASSWORD&button=login&__RequestVerificationToken=$REQUESTVERIFICATIONTOKEN"  https://jouw.postnl.nl/identity/Account/login`
+STATUSCODE=`curl --tlsv1.2 $PROXY -i -w "\n%{http_code}" -o $RESPFILE -k -s -b $COOKIEFILE -c $COOKIEFILE -L --header 'Connection: close' --data-urlencode "Username=$USERNAME" --data-urlencode "Password=$PASSWORD" --data-urlencode "ReturnUrl=$RETURNURL" --data-urlencode "__RequestVerificationToken=$REQUESTVERIFICATIONTOKEN" --data-urlencode 'button=login' https://jouw.postnl.nl/identity/Account/login`
+
 if [ ! $STATUSCODE = "200" ]
 then
 	echo "Error: Wrong responsecode $STATUSCODE received from Account/Login step 3 request. See $RESPFILE. Aborting."
@@ -242,26 +296,29 @@ then
 	exit 1
 fi
 
+
 # Generate CodeVerifier, stateValue and codeChallenge
 
 RANDOM=`dd bs=32 count=1 status=none </dev/urandom`
 CODEVERIFIER=`echo -n $RANDOM | hexdump -e '32/1 "%02x""\n"'`
-# echo "CodeVerifier: $CODEVERIFIER"
+debug_output "CodeVerifier: $CODEVERIFIER"
 TMP1=`echo -n "$CODEVERIFIER" | openssl dgst -sha256 -binary | openssl base64`
 CODECHALLENGE=`echo -n $TMP1 | sed -e 's/\+/-/g' -e 's/\//_/g' -e 's/=//g'`
-# echo "codeChallenge: $CODECHALLENGE"
+debug_output "codeChallenge: $CODECHALLENGE"
 
 STATEVALUE=`dd bs=32 count=1 </dev/urandom status=none | hexdump -e '32/1 "%02x""\n"'`
-# echo "stateValue: $STATEVALUE"
+debug_output "stateValue: $STATEVALUE"
 
 
 # Step 4
+debug_output ""
+debug_output "Start step 4"
 
 RESPFILE="$TMPDIR/postnl-resp4.tmp"
 removeFile $RESPFILE
 
-OUTPUTCURL=`curl $PROXY -i -o $RESPFILE -k -b $COOKIEFILE -c $COOKIEFILE -s -L -w %{url_effective} "https://jouw.postnl.nl/identity/connect/authorize?client_id=pwb-web&audience=poa-profiles-api&scope=openid%20profile%20email%20poa-profiles-api%20pwb-web-api&response_type=code&code_challenge_method=S256&code_challenge=$CODECHALLENGE&prompt=none&state=$STATEVALUE&redirect_uri=https://jouw.postnl.nl/silent-renew.html&ui_locales=nl_NL"`
-#echo "Output identity/connect/authorize: $OUTPUTCURL"
+OUTPUTCURL=`curl --tlsv1.2 $PROXY -i -o $RESPFILE -k -b $COOKIEFILE -c $COOKIEFILE -s -L -w %{url_effective} --header 'Connection: close' --request GET "https://jouw.postnl.nl/identity/connect/authorize?client_id=pwb-web&audience=poa-profiles-api&scope=openid%20profile%20email%20poa-profiles-api%20pwb-web-api&response_type=code&code_challenge_method=S256&code_challenge=$CODECHALLENGE&prompt=none&state=$STATEVALUE&redirect_uri=https://jouw.postnl.nl/silent-renew.html&ui_locales=nl_NL"`
+#debug_output "Output identity/connect/authorize: $OUTPUTCURL"
 
 if ! echo $OUTPUTCURL | grep -q "^https://jouw.postnl.nl"
 then
@@ -278,15 +335,17 @@ then
 fi
 
 CODE=`echo -n $OUTPUTCURL | sed 's/.*code=\(.*\)\&scope.*/\1/'`
-# echo "Code: $CODE"
+debug_output "Code: $CODE"
 
 
 # Step 5
+debug_output ""
+debug_output "Start step 5"
 
 RESPFILE="$TMPDIR/postnl-resp5.tmp"
 removeFile $RESPFILE
 
-STATUSCODE=`curl $PROXY -w "\n%{http_code}" -k -o $RESPFILE -b $COOKIEFILE -c $COOKIEFILE -s -d "grant_type=authorization_code&client_id=pwb-web&code=$CODE&code_verifier=$CODEVERIFIER&redirect_uri=https%3A%2F%2Fjouw.postnl.nl%2Fsilent-renew.html"  https://jouw.postnl.nl/identity/connect/token`
+STATUSCODE=`curl --tlsv1.2 $PROXY -w "\n%{http_code}" -k -o $RESPFILE -b $COOKIEFILE -c $COOKIEFILE -s -d "grant_type=authorization_code&client_id=pwb-web&code=$CODE&code_verifier=$CODEVERIFIER&redirect_uri=https%3A%2F%2Fjouw.postnl.nl%2Fsilent-renew.html"  https://jouw.postnl.nl/identity/connect/token`
 if [ ! $STATUSCODE = "200" ]
 then
 	echo "Error: Wrong responsecode $STATUSCODE received from connect/token request. See $RESPFILE. Aborting."
@@ -304,15 +363,25 @@ then
 fi
 
 ACCESSCODE=`grep -o '"access_token":"[^"]*' $RESPFILE | grep -o '[^"]*$'`
-# echo "Access_code: $ACCESSCODE"
+debug_output "Access_code: $ACCESSCODE"
 echo -n $ACCESSCODE > $ACCESSCODEFILE
 
 # Step 6
+debug_output ""
+debug_output "Start step 6"
 
 readInbox $ACCESSCODEFILE
 
 
 #Remove response files.
-rm $TMPDIR/postnl-resp?.tmp
+if ! $DEBUG
+then
+	rm $TMPDIR/postnl-resp?.tmp
+fi
 
 # Done
+
+
+
+
+
